# Gather qmm batched kernel and refactoring of quantized