… benchmarks/python/gather_qmm_bench.py
… mlx/ops.cpp
… mlx/ops.h
… mlx/primitives.cpp
… mlx/primitives.h
… python/src/ops.cpp
… python/tests/test_quantized.py