This is part 2 of the #2040 kind of. It implements the same kernel but for quantized mm and refactors `quantized.cpp`. It also sets up `quantized.h` for a refactor of the batched kernel (merge `qmm_n` and `qmm_t` and more). 

In the process of refactoring it fixes the following 2 bugs
- routing `gather_qmv` to `qmv_quad`
- assuming N >= 64 for `qvm`

In the microbenchmark we see similar speedups as for #2040. A little less impressive because `qmv` is faster and we have to skip less data to move from matrix to matrix. I am adding some real world MoE prompt processing speedups below

| Model | Prompt size | Before unsorted | Before sorted | Now sorted |
| --- | --- | --- | --- | --- |
| Mixtral 8x7B | ~500 | 171 tps | 189 tps | 590 tps |
| Mixtral 8x7B | ~6000 | 179 tps | 196 tps | 681 tps |
| Qwen 1.5 2.7B | ~500 | 1071 tps | 1239 tps | 2213 tps |
| Qwen 1.5 2.7B | ~6000 | 1333 tps | 1532 tps | 3360 tps |
| Llama 4 Scout | ~500 | 244 tps | 253 tps | 429 tps |
| Llama 4 Scout | ~6000 | 258 tps | 264 tps | 526 tps |