@angeloskath 2025-04-17
For DeepSeek with its 256 experts we have a smaller speedup for small prompts and we 're back at 2x when there are enough tokens to fill the experts.

| Model | Context | Before sorted | Now sorted |
| --- | --- | --- | --- |
| Deepseek V3 | ~450 | 112 tps | 154 tps |
| Deepseek V3 | ~2000 | 114 tps | 210 tps |

As expected I see no regression in non MoE models so I think this is good to merge.